move walk to render utils
