+++ core/src/plugins/render/utils/walk.ts
